+++ pom.xml
@@ -53,7 +53,7 @@
                             <transformers>
                                 <transformer
                                         implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                                    <mainClass>org.bozdgn.App</mainClass>
+                                    <mainClass>org.bozdgn.client.App</mainClass>
                                 </transformer>
                             </transformers>
                         </configuration>
@@ -74,7 +74,7 @@
                 <artifactId>javafx-maven-plugin</artifactId>
                 <version>0.0.8</version>
                 <configuration>
-                    <mainClass>org.bozdgn.App</mainClass>
+                    <mainClass>org.bozdgn.client.App</mainClass>
                 </configuration>
             </plugin>
         </plugins>
