+++ src/main/java/org/bozdgn/client/data/AdminPurchase.java
@@ -1,17 +1,27 @@
-package org.bozdgn.model;
+package org.bozdgn.client.data;
 
 import java.time.LocalDate;
 
-public class Purchase{
+public class AdminPurchase{
+    private String pid;
+    private int mid;
     private LocalDate date;
     private String repast;
     private String refectory;
 
-    public Purchase(LocalDate date, String repast, String refectory){
+    public AdminPurchase(String pid, int mid, LocalDate date, String repast, String refectory){
+        this.pid = pid;
+        this.mid = mid;
         this.date = date;
         this.repast = repast;
         this.refectory = refectory;
     }
+
+    public String getPid(){ return pid; }
+    public void setPid(String pid){ this.pid = pid; }
+
+    public int getMid(){ return mid; }
+    public void setMid(int mid){ this.mid = mid; }
 
     public LocalDate getDate(){ return date; }
     public void setDate(LocalDate date){ this.date = date; }
