+++ src/main/java/org/bozdgn/service/Database.java
@@ -1,4 +1,4 @@
-package org.bozdgn;
+package org.bozdgn.service;
 
 import java.sql.*;
 import java.util.*;
